refactor: migrate to new Header component and update navigation styling
